Moved to top level UnsupportedServerConfig, other PR feedback
